requirements.txt
streamlit==1.53.1
pandas>=2.3.3
google-genai>=1.60.0
pydeck>=0.9.1
plotly>=6.5.2
selenium>=4.40.0
webdriver-manager>=4.0.2
beautifulsoup4>=4.14.3
google-generativeai>=0.8.6
psutil>=7.2.1
requests>=2.32.5
schedule>=1.2.2
python-dotenv>=1.2.1
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
